refactor #5: directory reorder
