+++ src/App.js
@@ -4,8 +4,8 @@ import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 // importamos los componentes creados
 import Navbar from "./components/layout/Navbar";
 import Footer from "./components/layout/Footer";
-import Home from "./components/pages/partners/Home";
-import Items from "./components/pages/partners/Items";
+import Home from "./pages/partners/Home";
+import Items from "./pages/partners/Items";
 
 var navLinks= [
     {
