+++ src/pages/partners/Home.js
